requirements.txt
Jinja2
requests
dnspython
Click
python-dotenv
jsonschema
appdirs